Move it to a different class so it doesn't break fragile dynamic policy class loading from module.
