+++ st2actions/st2actions/policies/concurrency.py
@@ -16,34 +16,14 @@
 from st2common.constants import action as action_constants
 from st2common import log as logging
 from st2common.persistence import action as action_access
-from st2common.policies import base
+from st2common.policies.concurrency import BaseConcurrencyApplicator
 from st2common.services import action as action_service
-from st2common.services import coordination
 
 __all__ = [
-    'BaseConcurrencyApplicator',
     'ConcurrencyApplicator'
 ]
 
 LOG = logging.getLogger(__name__)
-
-
-class BaseConcurrencyApplicator(base.ResourcePolicyApplicator):
-    def __init__(self, policy_ref, policy_type, threshold=0, action='delay'):
-        super(BaseConcurrencyApplicator, self).__init__(policy_ref=policy_ref,
-                                                        policy_type=policy_type)
-        self.threshold = threshold
-        self.policy_action = action
-
-        self.coordinator = coordination.get_coordinator()
-
-    def _get_status_for_policy_action(self, policy_action):
-        if policy_action == 'delay':
-            status = action_constants.LIVEACTION_STATUS_DELAYED
-        elif policy_action == 'cancel':
-            status = action_constants.LIVEACTION_STATUS_CANCELED
-
-        return status
 
 
 class ConcurrencyApplicator(BaseConcurrencyApplicator):
@@ -129,4 +109,3 @@ class ConcurrencyApplicator(BaseConcurrencyApplicator):
             self._apply_after(target)
 
         return target
-
